backend/requirements.txt
# requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
requests>=2.31.0
pyyaml>=6.0.1
python-multipart>=0.0.6
ujson>=5.8.0
httpx>=0.24.1